Refactored website to have HTML generated by Template::Toolkit (TT2)
TT2 files live in the src/ directory, and *.tt2 files are processed to
produce matching ../*.html.

Also took the opportunity to change the CSS and visual appearance of the
site. The HTML content "looks" the same, but has changed some internal structure
because it is now being generated by TT2